chore: Rename service-discovery orchestrator
